+++ packages/terra-application-navigation/tests/wdio/application-navigation-spec.js
@@ -317,3 +317,25 @@ Terra.describeViewports('ApplicationNavigation Responsive', ['small', 'medium', 
     Terra.it.validatesElement({ selector: '#root' });
   });
 });
+
+Terra.describeViewports('ApplicationNavigation', ['large', 'huge', 'enormous'], () => {
+  describe('Should open selected item when alert is accepted', () => {
+    before(() => {
+      browser.url('/#/raw/tests/terra-application-navigation/application-navigation/application-navigation-prompt-focus');
+      browser.waitForVisible('[aria-label="Management"]', 5000);
+      browser.click('[aria-label="Management"]');
+    });
+
+    it('Accept alert', () => {
+      browser.alertAccept();
+    });
+
+    Terra.it.validatesElement();
+
+    it('should focus content after a tab', () => {
+      browser.keys('Tab');
+    });
+
+    Terra.it.validatesElement('focus should be on content');
+  });
+});