+++ packages/terra-application-navigation/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Focus should be returned to `Tab` key clicked when `Navigation prompt` is dismissed.
+* Focus should be passed to `Content` when `Navigation prompt` is accepted.
 
 1.27.0 - (May 19, 2020)
 ------------------